requirements.txt
google-genai
google-cloud-discoveryengine
pypdf
python-docx
PyYAML
python-dotenv
numpy
